requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2